requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# OpenAI
openai==1.58.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6

# LangChain for RAG
langchain==0.3.14
langchain-openai==0.2.14
langchain-postgres==0.0.13

# Document Processing
pypdf2==3.0.1
python-docx==1.1.2
pdfplumber==0.11.4

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
aiofiles==24.1.0
